+++ views/canvas/admin.ejs
@@ -1,6 +1,6 @@
 <% layout('layout') %>
 
-  <form action="/game/admin" method="post" class="form-register" id="form-register">
+  <div class="form-register" id="form-register">
     <h1 class="h3 mb-3 font-weight-normal">Administration panel</h1>
 
     <div class="form-group">
@@ -19,6 +19,6 @@
 
     <button class="btn btn-lg btn-primary btn-block" type="submit" id="save">Save</button>
     <iframe src="/status" style="width:100%; height: 1030px; border: solid 1px #ccc; margin-top:20px;"></iframe>
-  </form>
+  </div>
 
-  <script type="module" src="/js/user/admin.js"></script>
+  <script type="module" src="/js/canvas/admin.js"></script>
